requirements.txt
google-cloud-bigquery
google-auth
requests
pandas
jupyter